+++ app/Http/Controllers/YouthReferenceController.php
@@ -34,6 +34,8 @@ class YouthReferenceController extends Controller
      */
     public function getList(Request $request): JsonResponse
     {
+        $this->authorize('view', YouthReference::class);
+
         $filter = $this->referenceService->filterValidator($request)->validate();
         $response = $this->referenceService->getReferenceList($filter, $this->startTime);
         return Response::json($response, ResponseAlias::HTTP_OK);
@@ -48,9 +50,11 @@ class YouthReferenceController extends Controller
      */
     public function read(int $id): JsonResponse
     {
-        $data = $this->referenceService->getOneReference($id);
+        $reference = $this->referenceService->getOneReference($id);
+        $this->authorize('view', $reference);
+
         $response = [
-            "data" => $data,
+            "data" => $reference,
             "_response_status" => [
                 "success" => true,
                 "code" => ResponseAlias::HTTP_OK,
@@ -70,6 +74,8 @@ class YouthReferenceController extends Controller
      */
     public function store(Request $request): JsonResponse
     {
+        $this->authorize('create', YouthReference::class);
+
         if (!$request->filled('youth_id')) {
             $youthId = Auth::id();
             $request->offsetSet('youth_id', $youthId);
@@ -100,6 +106,9 @@ class YouthReferenceController extends Controller
     public function update(Request $request, int $id): JsonResponse
     {
         $reference = YouthReference::findOrFail($id);
+
+        $this->authorize('update', $reference);
+
         if (!$request->filled('youth_id')) {
             $youthId = Auth::id();
             $request->offsetSet('youth_id', $youthId);
@@ -127,6 +136,8 @@ class YouthReferenceController extends Controller
     public function destroy(int $id): JsonResponse
     {
         $reference = YouthReference::findOrFail($id);
+        $this->authorize('delete', $reference);
+
         $this->referenceService->destroy($reference);
         $response = [
             '_response_status' => [
